refactor(ir): ✓ move regexp-runtime-contract below the IR (#3113 S2)
`STANDALONE_REGEXP_CARRIER_TEST_HELPER` is shared vocabulary — a single
dependency-free constant naming a Wasm helper that the IR front-end and the
legacy codegen must agree on. Its own doc comment says so ("shared by legacy
codegen and IR"), but it lived in src/codegen/, which forced src/ir/ to
import upward: from-ast.ts and integration.ts both reached for it.

Same shape as the js-tag.ts move in slice 1, and the same criterion: the
module has ZERO imports, so relocating it cannot introduce a cycle or drag
a dependency along. src/codegen/regexp-standalone.ts now imports it
down-stack (../ir/), which is the correct direction.

Pure relocation, no logic change:
  - prove-emit-identity check → IDENTICAL, all 60 (file,target) emits
  - ts7 typecheck clean
  - ir-layering ratchet 96 → 94, banked in the baseline
