+++ scripts/ir-layering-baseline.json
@@ -1,6 +1,6 @@
 {
-  "generated": "2026-08-21T10:48:42.598Z",
-  "total": 96,
+  "generated": "2026-08-21T11:00:59.142Z",
+  "total": 94,
   "files": {
     "ir/analysis/i32-slots.ts": 2,
     "ir/backend/linear-integration.ts": 1,
@@ -11,9 +11,9 @@
     "ir/dynamic-number-lowering.ts": 3,
     "ir/fixed-literal-loop-proof.ts": 1,
     "ir/fmod-selection.ts": 2,
-    "ir/from-ast.ts": 7,
+    "ir/from-ast.ts": 6,
     "ir/i32-pure-bitwise.ts": 2,
-    "ir/integration.ts": 42,
+    "ir/integration.ts": 41,
     "ir/math-runtime-providers.ts": 3,
     "ir/number-to-string-provider.ts": 6,
     "ir/prepared-callable-resolution.ts": 3,
